Fix replacing file by url
