+++ lib/src/resources/app_assets.dart
@@ -2,7 +2,6 @@ final class AppAssets {
   const AppAssets._();
 
   static const imagesLogo = 'assets/images/logo.png';
-  static const imagesUserProfile = 'assets/images/user_profile.jpeg';
   static const imagesLogoBantubeat = 'assets/images/logo_bantubeat.png';
   static const imagesFeatlinkLogo = 'assets/images/featlink_logo.png';
   static const imagesBeat = 'assets/images/beat.png';
