--- lib/src/screens/profile_creation/profile_create_user_profile_screen.dart
@@ -1,204 +1,0 @@
-import 'package:easy_localization/easy_localization.dart';
-import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:saloonprived_app/generated/locale_keys.g.dart';
-import 'package:saloonprived_app/src/resources/app_assets.dart';
-
-class ProfileCreateUserProfileScreen extends StatelessWidget {
-  const ProfileCreateUserProfileScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      body: Center(
-        child: SizedBox(
-          height: MediaQuery.of(context).size.height * 0.6 + 30,
-          child: Stack(
-            children: [
-              Container(
-                width: MediaQuery.of(context).size.width * 0.8,
-                height: MediaQuery.of(context).size.height * 0.6,
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 20.0,
-                  vertical: 40.0,
-                ),
-                decoration: BoxDecoration(
-                  color: const Color.fromRGBO(243, 244, 246, 1),
-                  borderRadius: BorderRadius.circular(20),
-                ),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text(
-                      LocaleKeys.profile_create_user_profile_screen_title.tr(),
-                      style: GoogleFonts.inter(
-                        color: Colors.black,
-                        fontSize: 12,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                    Column(
-                      children: [
-                        Container(
-                          decoration: BoxDecoration(
-                            shape: BoxShape.circle,
-                            border: Border.all(
-                              color: const Color.fromRGBO(252, 204, 55, 1),
-                              width: 2,
-                            ),
-                          ),
-                          child: ClipOval(
-                            child: Image.asset(
-                              AppAssets.imagesUserProfile,
-                              width: 100,
-                              height: 100,
-                              fit: BoxFit.cover,
-                            ),
-                          ),
-                        ),
-                        Text(
-                          'Afrika Kemie',
-                          style: GoogleFonts.inter(
-                            color: Colors.black,
-                            fontSize: 12,
-                            fontWeight: FontWeight.w700,
-                          ),
-                        ),
-                        const SizedBox(
-                          height: 10,
-                        ),
-                        Wrap(
-                          alignment: WrapAlignment.spaceEvenly,
-                          children: [
-                            RichText(
-                              text: TextSpan(
-                                children: [
-                                  TextSpan(
-                                    text: '10k',
-                                    style: GoogleFonts.inter(
-                                      color: Colors.black,
-                                      fontSize: 12,
-                                      fontWeight: FontWeight.w700,
-                                    ),
-                                  ),
-                                  TextSpan(
-                                    text: LocaleKeys
-                                        .profile_create_user_profile_screen_subscriptions_count
-                                        .tr(),
-                                    style: GoogleFonts.inter(
-                                      color: Colors.black,
-                                      fontSize: 12,
-                                      fontWeight: FontWeight.w400,
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 15,
-                            ),
-                            RichText(
-                              text: TextSpan(
-                                children: [
-                                  TextSpan(
-                                    text: '10k',
-                                    style: GoogleFonts.inter(
-                                      color: Colors.black,
-                                      fontSize: 12,
-                                      fontWeight: FontWeight.w700,
-                                    ),
-                                  ),
-                                  TextSpan(
-                                    text: LocaleKeys
-                                        .profile_create_user_profile_screen_followers_count
-                                        .tr(),
-                                    style: GoogleFonts.inter(
-                                      color: Colors.black,
-                                      fontSize: 12,
-                                      fontWeight: FontWeight.w400,
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ],
-                        ),
-                        const SizedBox(
-                          height: 25,
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-              Positioned(
-                bottom: 20,
-                right: 10,
-                left: 10,
-                child: Wrap(
-                  alignment: WrapAlignment.spaceAround,
-                  children: [
-                    ElevatedButton(
-                      onPressed: () {
-                        Navigator.pop(context);
-                      },
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors
-                            .black, // Couleur noire pour le bouton "Annuler"
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(20),
-                        ),
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 25,
-                          vertical: 15,
-                        ),
-                      ),
-                      child: Text(
-                        LocaleKeys
-                            .profile_create_user_profile_screen_cancel_button
-                            .tr(),
-                        style: GoogleFonts.inter(
-                          color: Colors.white,
-                          fontSize: 12,
-                          fontWeight: FontWeight.w700,
-                        ),
-                      ),
-                    ),
-                    ElevatedButton(
-                      onPressed: () {
-                        // Action pour ajouter
-                      },
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: const Color(
-                          0xFFFCCC37,
-                        ), // Couleur orange pour le bouton "Ajouter"
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(20),
-                        ),
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 25,
-                          vertical: 15,
-                        ),
-                      ),
-                      child: Text(
-                        LocaleKeys.profile_create_user_profile_screen_add_button
-                            .tr(),
-                        style: GoogleFonts.inter(
-                          color: Colors.white,
-                          fontSize: 12,
-                          fontWeight: FontWeight.w700,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              // Bouton "Ajouter" en dehors de la bordure droite
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
